add check for same user to delete post

+++ src/api/public/posts/posts.service.ts
@@ -80,7 +80,13 @@ export class PostsService {
     if (!post) return null
     return post
   }
-  async deletePost(id: string) : Promise<boolean> {
+  async deletePost(id: string, token : string) : Promise<boolean> {
+    const post : PostModel = await this.postsRepository.getPost(id)
+    if (!post) throw new NotFoundException()
+    const userId : string = await this.jwtService.getUserIdByToken(token)
+    const blog : BlogModel = await this.blogsRepository.getFullBlog(post.blogId)
+    if (!blog) throw new NotFoundException()
+    if (blog.blogOwnerInfo.userId !== userId) throw new ForbiddenException()
     return await this.postsRepository.deletePost(id)
   }
   async createPost(post: PostsDto, token : string) : Promise <PostModel | null>{
